Postfix for FB8-208 (Configure Jenkins job for fb-mysql-8.0.13 branch) (#97)
https://jira.percona.com/browse/FB8-208

Fixed the way how 'zstd' library is installed in 'docker/install-deps'.

For RHEL and CentOS (yum-based distros) we now install only 'libzstd-devel'
(available in EPEL). 'libzstd' will be installed as a depencency.

For Ubuntu (apt-based):
* we install nothing for Trusty (14.X) and lower;
* we install 'libzstd1-dev' for Xenial (16.X);
* we install 'libzstd-dev' for Bionic (18.X) and higher.
'libzstd' / 'libzstd1' will be installed automatically as a depencency.

For Debian (apt-based):
* we install nothing for Jessie (8.X) and lower;
* we install 'libzstd-dev' for Stretch (9.X).
'libzstd' / 'libzstd1' will be installed automatically as a depencency.

Also, set proper shell in the 'docker/prepare-docker' script.