PS-10120: Fix issues with broken "sources" directory
`sources` are downloaded only for Build worker that becomes MTR worker 1 later.
For other workers `sources` shouldn't exist.
But in logs we have seen that it's possible that a Hetzner instance was not cleaned from a previous run.
```
fatal: detected dubious ownership in repository at '/mnt/jenkins/workspace/percona-server-9.x-pipeline-parallel-mtr/sources'
To add an exception for this directory, call:

	git config --global --add safe.directory /mnt/jenkins/workspace/percona-server-9.x-pipeline-parallel-mtr/sources
```
This patch checks if `sources` exists and is a valid git repo with a proper permissions.
In case repo is not valid it's removed as it probably comes from a previous run which was not cleaned properly.