PS-10223: Add ARM platforms to Jenkins
https://perconadev.atlassian.net/browse/PS-10223

Extended prepare-ps-build-docker pipeline and supporting script so it
can now prepare ps-build docker images, which are used for PS build and
test runs in Jenkins, for ARM64 platform.

These images are uploaded to ECR with -aarch64 suffix in their tag so
they can be correctly identified when used from pipeline-parallel-mtr-*.*
pipelines.

Added new ARCH parameter to pipeline-parallel-mtr-*.* pipelines, which
allows to choose whether we want to build and test on x86_64 (default)
or aarch64 (ARM64) architecture. Adjusted scripts for building server
and running tests to take this parameter/the fact that they can be
run on ARM64 into account.

Added new ARCH axis to matrix parameter of param-parallel-mtr-*.* job.
Combinations using aarch64 (ARM64) architecture are not enabled by
default.